backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]>=0.31.1
SQLAlchemy==2.0.27
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.5
email-validator==2.1.0.post1
python-multipart==0.0.20
pydantic-settings==2.9.1
python-json-logger==3.3.0
motor==3.7.0
dnspython==2.7.0
ecdsa>=0.13.3,!=0.15
starlette>=0.40.0,<0.47.0
pydantic>=1.7.4,!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0
typing-extensions>=4.8.0
greenlet!=0.4.17
rsa>=4.0,!=4.1.1,!=4.4,<5.0
itsdangerous==2.1.2